Client/Playlist.cpp: Remove SSL config definition
Well, this just isn't a Qt class on Windows for some reason.  This might
or might not break SSL functionality while downloading remote URLs over HTTPS,
but we do not seem to support streamed audio in the mainline game yet, only in
Audiotest, so we do not need to worry about it that much anyways, yet.