Puyolib: Basic console debug (#45)
* Puyolib: Add DebugLog

This class will allow a further integration of debugging capability into
Puyolib by allowing it to link to a debug output of any supported client.

This change and the following ones might be breaking so please check
compatiblity with all clients before merging.

* Puyolib: Force Game::Game to take a DebugLog

This might seem heavy-handed but is the only feasible measure I could figure out
that would allow for debugging all of the Puyolib code, including the constructor.

* Client: Add DebugLog support

This implements basic debugging of Puyolib inside of Client. Currently prints
anything that gets generated regardless of its state or flags.

* Puyolib: Add documentation to the DebugLog class

* Puyolib: Fix memory leaking of DebugLog

We were previously leaking memory per Game object destroyed.

* Puyolib: Code review refactoring

* Puyolib: Fix debugString in debugMode

Fixes the ancient debug status text for onscreen use while in debug mode.

* Puyolib: Implement and fix current Puyolib debugging

Restores the debug status text functionality and gives more documentation
to Puyolib debug mode. For DebugLog, debug display is up to interpretation
of the client.

* ClientNG: Add support for Puyolib/DebugLog

It's possible it's leaking memory (but it's an unique pointer so it should get
GC'd, hopefully...) but it works!

* Puyolib, Client: condition DebugLog in Client on Puyolib::debugMode

* Puyolib: Avoid Windows inserting its own macros

* Puyolib: fix previous commit