Puyolib: Asset management unification and source overlaying (#52)
* Puyolib: Add DebugLog

This class will allow a further integration of debugging capability into
Puyolib by allowing it to link to a debug output of any supported client.

This change and the following ones might be breaking so please check
compatiblity with all clients before merging.

* Puyolib: Force Game::Game to take a DebugLog

This might seem heavy-handed but is the only feasible measure I could figure out
that would allow for debugging all of the Puyolib code, including the constructor.

* Client: Add DebugLog support

This implements basic debugging of Puyolib inside of Client. Currently prints
anything that gets generated regardless of its state or flags.

* Puyolib: Add documentation to the DebugLog class

* Puyolib: Fix memory leaking of DebugLog

We were previously leaking memory per Game object destroyed.

* Puyolib: Code review refactoring

* Puyolib: Fix debugString in debugMode

Fixes the ancient debug status text for onscreen use while in debug mode.

* Puyolib: Implement and fix current Puyolib debugging

Restores the debug status text functionality and gives more documentation
to Puyolib debug mode. For DebugLog, debug display is up to interpretation
of the client.

* ClientNG: Add support for Puyolib/DebugLog

It's possible it's leaking memory (but it's an unique pointer so it should get
GC'd, hopefully...) but it works!

* Puyolib, Client: condition DebugLog in Client on Puyolib::debugMode

* Puyolib: Avoid Windows inserting its own macros

* Puyolib: fix previous commit

* Puyolib: Force the usake of Base directory for voice assets
Previously this wasn't being done and was causing issues in portable mode

* ClientNG: implement basic audio read in Scenes/Game
No caching is implemented yet.

* ClientNG: Add caching to audio files in Scenes/Game

* Fix previous commit

* ClientNG: Link Renderer filters to Game Frontend

* ClientNG/Scenes/Game: Port Music Playlist primitives

* ClientNG: Notify user of errors on unsuccessful loadImage

* Puyolib: Add a skeleton for abstracting asset importing

* Puyolib: implement FileAssetBundle filename convertor

* Puyolib: Add error checking in Audiolib, file listings

* Client/Frontend: Add a sound load test

* Puyolib: syntax changes

* Puyolib: Implement AssetManager
Accidentally also implements basic support into ClientNG 😅

* Client: support AssetManager

* Puyolib: fix AssetBundles and AssetManager

* Puyolib: remove redundant macro

* Puyolib: link ppvs::Game to AssetManager

* Puyolib: introduce AssetManager support to the Player class

* Puyolib: add AssetBundle contents iterators

* Puyo: support animations with AssetManager

* Puyolib: allow cloning of AssetManager and AssetBundle

* Puyolib: Clean asset loading

* Client: Make use of AssetManager

* Client: fix file-browsing bugs with AssetManager
I broke my repo here so it's possible there will be some nonsense changes in this commit or the next one, revert if needed.

* Client: refactor MainWindow

* Puyolib: Resolve MSVC complaining about implicit string conversions

* Update Puyolib/CMakeLists.txt

Co-authored-by: jchv <john@jchw.io>

* Update ClientNG/Scenes/Game/Frontend.h

Co-authored-by: jchv <john@jchw.io>

* Code review changes

* Puyolib: modern nullptr definitions

* Update ClientNG/Scenes/Game/MusicPlayer.h

Co-authored-by: jchv <john@jchw.io>

* Update Client/gamewidgetgl.cpp

Co-authored-by: jchv <john@jchw.io>

* Update Puyolib/AssetBundle.cpp

Co-authored-by: jchv <john@jchw.io>

* Update Puyolib/Game.h

Co-authored-by: jchv <john@jchw.io>

* Apply suggestions from code review

Co-authored-by: jchv <john@jchw.io>

* Code review: rename "generateAM"

* Puyolib: Move from text tokens to enum classes

* Minor reformatting

---------

Co-authored-by: jchv <john@jchw.io>